LRQA-40401 Change "TestClass" & "TestMethod" classes into inner classes
