+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/TestClassGroup.java
@@ -16,15 +16,72 @@ package com.liferay.jenkins.results.parser;
 
 import java.io.File;
 
+import java.util.ArrayList;
 import java.util.List;
 
 /**
  * @author Michael Hashimoto
  */
-public interface TestClassGroup {
+public abstract class TestClassGroup {
 
-	public List<TestClass> getTestClasses();
+	public abstract List<TestClass> getTestClasses();
 
-	public List<File> getTestClassFiles();
+	public abstract List<File> getTestClassFiles();
+
+	protected static class TestClass implements Comparable<TestClass> {
+
+		@Override
+		public int compareTo(TestClass testClass) {
+			if (testClass == null) {
+				throw new NullPointerException("test class is NULL");
+			}
+
+			return _file.compareTo(testClass.getFile());
+		}
+
+		protected TestClass(File file) {
+			_file = file;
+		}
+
+		protected void addTestMethod(String methodName) {
+			addTestMethod(new TestMethod(methodName, this));
+		}
+
+		protected void addTestMethod(TestMethod testMethod) {
+			_testMethods.add(testMethod);
+		}
+
+		protected File getFile() {
+			return _file;
+		}
+
+		protected List<TestMethod> getTestMethods() {
+			return _testMethods;
+		}
+
+		private final File _file;
+		private final List<TestMethod> _testMethods = new ArrayList<>();
+
+	}
+
+	protected static class TestMethod {
+
+		protected TestMethod(String name, TestClass testClass) {
+			_name = name;
+			_testClass = testClass;
+		}
+
+		protected String getName() {
+			return _name;
+		}
+
+		protected TestClass getTestClass() {
+			return _testClass;
+		}
+
+		private final String _name;
+		private final TestClass _testClass;
+
+	}
 
 }
